move commadns into separate module and put their options next to them
