+++ src/commands/dry_run.rs
@@ -18,8 +18,35 @@
 
 use pallet_election_provider_multi_phase::RawSolution;
 
-use crate::{epm, error::Error, opt::DryRunConfig, prelude::*, signer::Signer, static_types};
+use crate::{epm, error::Error, opt::Solver, prelude::*, signer::Signer, static_types};
+use clap::Parser;
 use codec::Encode;
+
+#[derive(Debug, Clone, Parser)]
+#[cfg_attr(test, derive(PartialEq))]
+pub struct DryRunConfig {
+	/// The block hash at which scraping happens. If none is provided, the latest head is used.
+	#[clap(long)]
+	pub at: Option<Hash>,
+
+	/// The solver algorithm to use.
+	#[clap(subcommand)]
+	pub solver: Solver,
+
+	/// Force create a new snapshot, else expect one to exist onchain.
+	#[clap(long)]
+	pub force_snapshot: bool,
+
+	/// The path to a file containing the seed of the account. If the file is not found, the seed is
+	/// used as-is.
+	///
+	/// Can also be provided via the `SEED` environment variable.
+	///
+	/// WARNING: Don't use an account with a large stash for this. Based on how the bot is
+	/// configured, it might re-try and lose funds through transaction fees/deposits.
+	#[clap(long, short, env = "SEED")]
+	pub seed_or_path: String,
+}
 
 pub async fn dry_run_cmd<T>(api: SubxtClient, config: DryRunConfig) -> Result<(), Error>
 where
