+++ src/lib.rs
@@ -16,12 +16,10 @@
 
 #![allow(dead_code)]
 
-pub mod dry_run;
-pub mod emergency_solution;
+pub mod commands;
 pub mod epm;
 pub mod error;
 pub mod helpers;
-pub mod monitor;
 pub mod opt;
 pub mod prelude;
 pub mod prometheus;
